+++ IcingTaskManager@json/locale.sh
@@ -1,11 +1,13 @@
 #!/bin/sh
 
-cd "files/locale"
+cd "files/IcingTaskManager@json"
+cinnamon-json-makepot --js ./po/default.pot
+cd "po"
+
 for f in *.po
 do
- filebase=$(basename "$f")
- filename="${filebase%.*}"
- cp "./mo/$filename.mo" "$HOME/.local/share/locale/$filename/LC_MESSAGES/IcingTaskManager@json.mo"
- echo "Installed new languages file: $filename.mo to local."
+ msgmerge -U $f ./default.pot
+ echo "Updated $f with new definitions"
+ rm "$f~"
 done